+++ src/main/java/de/johni0702/minecraft/bobby/mixin/sodium/SodiumChunkManagerMixin.java
@@ -1,132 +1,45 @@
 package de.johni0702.minecraft.bobby.mixin.sodium;
 
-import de.johni0702.minecraft.bobby.FakeChunkManager;
-import de.johni0702.minecraft.bobby.FakeChunkStorage;
-import de.johni0702.minecraft.bobby.mixin.ClientChunkManagerMixin;
+import de.johni0702.minecraft.bobby.compat.IChunkStatusListener;
+import de.johni0702.minecraft.bobby.compat.sodium.SodiumChunkStatusListener;
+import de.johni0702.minecraft.bobby.ext.ClientChunkManagerExt;
 import me.jellysquid.mods.sodium.client.world.ChunkStatusListener;
-import me.jellysquid.mods.sodium.client.world.SodiumChunkManager;
-import net.minecraft.client.world.ClientWorld;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.network.PacketByteBuf;
-import net.minecraft.world.biome.source.BiomeArray;
-import net.minecraft.world.chunk.ChunkStatus;
-import net.minecraft.world.chunk.WorldChunk;
-import org.spongepowered.asm.mixin.Final;
+import net.minecraft.client.world.ClientChunkManager;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(SodiumChunkManager.class)
-public abstract class SodiumChunkManagerMixin extends ClientChunkManagerMixin {
-    @Shadow(remap = false) @Final private WorldChunk emptyChunk;
-    @Shadow(remap = false) private ChunkStatusListener listener;
-    @Shadow public abstract WorldChunk getChunk(int x, int z, ChunkStatus status, boolean create);
+@Mixin(value = ClientChunkManager.class, priority = 1010) // higher than our normal one
+public abstract class SodiumChunkManagerMixin implements ClientChunkManagerExt {
+    /* Shadows the one in Sodium's Mixin */
+    private ChunkStatusListener listener;
 
+    private SodiumChunkStatusListener wrappedListener;
     private ChunkStatusListener suppressedListener;
 
     @Override
-    protected FakeChunkManager createBobbyChunkManager(ClientWorld world) {
-        return new FakeChunkManager.Sodium(world, (SodiumChunkManager) (Object) this);
+    public IChunkStatusListener bobby_getListener() {
+        if (listener == null) {
+            return null;
+        }
+        if (wrappedListener == null || wrappedListener.delegate != listener) {
+            wrappedListener = new SodiumChunkStatusListener(listener);
+        }
+        return wrappedListener;
     }
 
-    @Inject(method = "getChunk", at = @At("RETURN"), cancellable = true)
-    private void bobbyGetChunk(int x, int z, ChunkStatus chunkStatus, boolean orEmpty, CallbackInfoReturnable<WorldChunk> ci) {
-        // Did we find a live chunk?
-        if (ci.getReturnValue() != (orEmpty ? emptyChunk : null)) {
-            return;
-        }
-
-        if (bobbyChunkManager == null) {
-            return;
-        }
-
-        // Otherwise, see if we've got one
-        WorldChunk chunk = bobbyChunkManager.getChunk(x, z);
-        if (chunk != null) {
-            ci.setReturnValue(chunk);
-        }
-    }
-
-    @Inject(method = "loadChunkFromPacket", at = @At("HEAD"))
-    private void bobbyUnloadFakeChunk(int x, int z, BiomeArray biomes, PacketByteBuf buf, CompoundTag tag, int verticalStripBitmask, boolean complete, CallbackInfoReturnable<WorldChunk> cir) {
-        if (bobbyChunkManager == null) {
-            return;
-        }
-
-        if (bobbyChunkManager.getChunk(x, z) != null) {
-            // We'll be replacing a fake chunk with a real one.
-            // Suppress the chunk status listener so it does
-            // get removed before its re-rendered.
-            suppressedListener = listener;
-            listener = null;
-        }
-
-        // This needs to be called unconditionally because even if there is no chunk loaded at the moment,
-        // we might already have one queued which we need to cancel as otherwise it will overwrite the real one later.
-        bobbyChunkManager.unload(x, z, true);
-    }
-
-    @Inject(method = "loadChunkFromPacket", at = @At("RETURN"))
-    private void bobbyFakeChunkReplaced(int x, int z, BiomeArray biomes, PacketByteBuf buf, CompoundTag tag, int verticalStripBitmask, boolean complete, CallbackInfoReturnable<WorldChunk> cir) {
-        if (suppressedListener != null) {
-            listener = suppressedListener;
-            suppressedListener = null;
-
-            // However, if we failed to load the chunk from the packet for whatever reason,
-            // we need to notify the listener that the chunk has indeed been unloaded.
-            if (getChunk(x, z, ChunkStatus.FULL, false) == null) {
-                listener.onChunkRemoved(x, z);
-            }
-        }
-    }
-
-    @Inject(method = "unload", at = @At("HEAD"))
-    private void bobbySaveChunk(int chunkX, int chunkZ, CallbackInfo ci) {
-        if (bobbyChunkManager == null) {
-            return;
-        }
-
-        WorldChunk chunk = getChunk(chunkX, chunkZ, ChunkStatus.FULL, false);
-        if (chunk == null) {
-            return;
-        }
-
-        // We'll be replacing a real chunk with a fake one.
-        // Suppress the chunk status listener so it does
-        // get removed before its re-rendered.
+    @Override
+    public void bobby_suppressListener() {
         suppressedListener = listener;
         listener = null;
-
-        FakeChunkStorage storage = bobbyChunkManager.getStorage();
-        CompoundTag tag = storage.serialize(chunk, getLightingProvider());
-        storage.save(chunk.getPos(), tag);
-        bobbyChunkReplacement = tag;
     }
 
-    @Inject(method = "unload", at = @At("RETURN"))
-    private void bobbyReplaceChunk(int chunkX, int chunkZ, CallbackInfo ci) {
-        CompoundTag tag = bobbyChunkReplacement;
-        bobbyChunkReplacement = null;
-        if (tag == null) {
-            return;
-        }
-        bobbyChunkManager.load(chunkX, chunkZ, tag, bobbyChunkManager.getStorage());
-
+    @Override
+    public IChunkStatusListener bobby_restoreListener() {
         if (suppressedListener != null) {
             listener = suppressedListener;
             suppressedListener = null;
+            return bobby_getListener();
+        } else {
+            return null;
         }
-    }
-
-    @Inject(method = "getDebugString", at = @At("RETURN"), cancellable = true)
-    private void bobbyDebugString(CallbackInfoReturnable<String> cir) {
-        if (bobbyChunkManager == null) {
-            return;
-        }
-
-        cir.setReturnValue(cir.getReturnValue() + " " + bobbyChunkManager.getDebugString());
     }
 }