+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.2.0-SNAPSHOT
 - Add option to disable loading of block entities in fake chunks, active by default
+- Fix maximum value of view distance overwrite slider to be 32 instead of 16 (#21)
 
 ### 1.1.4
 - Fix chunk flicker with Sodium when real chunk receives full update